Create cleaanTimestampMKV.ps1
Time stamp issue?  I forget.  Sometimes I download videos and they stutter.  If I check the fps, they're something weird like 23.94.
This script uses ffmpeg, mkvextract, and mkvmerge to isolate the video and audio track then remux them as an mkv with correct fps.
If you use it, be sure to change the location of the 3 command line apps within the first couple lines of the script.